# Minor performance improvement in `bootstrap_boot()`